chore: split configs into two
