+++ app/scripts/HorizontalTiledPlot.jsx
@@ -10,10 +10,7 @@ import HorizontalItem from './HorizontalItem';
 import SortableList from './SortableList';
 
 // Utils
-import { or, resetD3BrushStyle, sum } from './utils';
-
-// Configs
-import { IS_TRACK_RANGE_SELECTABLE } from './configs';
+import { or, resetD3BrushStyle, sum, IS_TRACK_RANGE_SELECTABLE } from './utils';
 
 // Styles
 import styles from '../styles/HorizontalTiledPlot.module.scss'; // eslint-disable-line no-unused-vars
