+++ app/scripts/configs/README.md
@@ -1,0 +1,3 @@
+# configs
+
+This folder contains constant values that determine the visual parameters and options of HiGlass tracks and views.
