+++ app/scripts/configs/index.js
@@ -11,13 +11,10 @@ export { THEME_LIGHT, THEME_DARK, THEME_DEFAULT } from './themes';
 export { default as GLOBALS } from './globals';
 export { default as DATATYPE_TO_TRACK_TYPE } from './datatype-to-track-type';
 export { default as HEATED_OBJECT_MAP } from './colormaps';
-export { default as IS_TRACK_RANGE_SELECTABLE } from './is-track-range-selectable';
-export { default as OPTIONS_INFO } from './options-info';
 export { default as TRACKS_INFO } from './tracks-info';
 export { default as TRACKS_INFO_BY_TYPE } from './tracks-info-by-type';
 export { default as POSITIONS_BY_DATATYPE } from './positions-by-datatype';
 export { default as DEFAULT_TRACKS_FOR_DATATYPE } from './default-tracks-for-datatype';
-export { default as AVAILABLE_FOR_PLUGINS } from './available-for-plugins';
 export { default as AVAILABLE_TRACK_TYPES } from './available-track-types';
 export {
   NUM_PRECOMP_SUBSETS_PER_1D_TTILE,
