+++ app/scripts/plugins/get-data-fetcher.js
