+++ app/scripts/plugins/index.js
@@ -1,0 +1,2 @@
+export { default as getDataFetcher } from './get-data-fetcher';
+export { default as AVAILABLE_FOR_PLUGINS } from './available-for-plugins';
