+++ app/scripts/utils/index.js
@@ -83,6 +83,7 @@ export { default as DenseDataExtrema1D } from './DenseDataExtrema1D';
 export { default as DenseDataExtrema2D } from './DenseDataExtrema2D';
 export { default as getAggregationFunction } from './get-aggregation-function';
 export { default as selectedItemsToSize } from './selected-items-to-size';
+export { default as IS_TRACK_RANGE_SELECTABLE } from './is-track-range-selectable';
 
 export {
   getTrackObjectFromHGC,
